fix(module:tabs) Fix tab size issue

+++ components/tabs/Tabs.razor.cs
@@ -228,7 +228,12 @@ namespace AntDesign
         private string _navListStyle;
 
         private string _operationClass = "ant-tabs-nav-operations ant-tabs-nav-operations-hidden";
+
+        private const string HiddenStyle = "visibility: hidden; order: 1;";
+
         private string _operationStyle = "visibility: hidden; order: 1;";
+        private string _firstAddButtonStyle = HiddenStyle;
+        private string _secondAddButtonStyle = HiddenStyle;
 
         private decimal _scrollOffset;
         private decimal _scrollListWidth;
@@ -258,7 +263,7 @@ namespace AntDesign
         private bool _shownDropdown;
         private bool _needUpdateScrollListPosition;
         private bool _retryingGetSize;
-
+        private bool _hasNewTab = false;
         protected override void OnInitialized()
         {
             base.OnInitialized();
@@ -318,8 +323,20 @@ namespace AntDesign
             _tabs.Add(tabPane);
             _needUpdateScrollListPosition = true;
             StateHasChanged();
+            if (_hasNewTab)
+            {
+                ActivatePane(tabPane.Key);
+                _hasNewTab = false;
+            }
         }
-
+        private async Task OnAddTab()
+        {
+            if (OnAddClick.HasDelegate)
+            {
+                _hasNewTab = true;
+                await OnAddClick.InvokeAsync(null);
+            }
+        }
         internal void Complete()
         {
             if (_afterFirstRender)
@@ -391,14 +408,22 @@ namespace AntDesign
             // if it is active, need to activiate the previous tab.
             if (_activeKey == tab.Key)
             {
-                Previous();
+                if (tab.TabIndex > 0)
+                {
+                    Previous();
+                }
+                else
+                {
+                    Next();
+                }
             }
 
             _shouldRender = true;
-            _needUpdateScrollListPosition = tab.TabIndex != _tabs.Count - 1; // only update scroll list position when active tab is not the last one
+            _needUpdateScrollListPosition = true; // only update scroll list position when active tab is not the last one
 
-            _tabs.Remove(tab);
+            _tabs.Remove(tab);            
             StateHasChanged();
+
         }
 
         internal async Task HandleTabClick(TabPane tabPane)
@@ -522,7 +547,7 @@ namespace AntDesign
             // Needs to be optimized to render only one tab instead all the tabs
             StateHasChanged();
 
-            _ = FocusAsync(_activeTab.TabBtnRef);
+            //_ = FocusAsync(_activeTab.TabBtnRef);
         }
 
         public override Task SetParametersAsync(ParameterView parameters)
@@ -611,6 +636,18 @@ namespace AntDesign
                     DomEventListener.AddExclusive<string>(_navListRef, "wheel", OnWheel, true);
                 }
             }
+            _firstAddButtonStyle = _secondAddButtonStyle = HiddenStyle;
+            if (HasAddButton)
+            {
+                if (_scrollListWidth <= _wrapperWidth)
+                {
+                    _firstAddButtonStyle = string.Empty;
+                }
+                else
+                {
+                    _secondAddButtonStyle = string.Empty;
+                }
+            }
         }
 
         private void OnWheel(string json)
@@ -696,7 +733,7 @@ namespace AntDesign
             {
                 _inkStyle = $"left: {_activeTabElement.OffsetLeft}px; width: {_activeTabElement.ClientWidth}px";
 
-                var additionalWidth = HasAddButton ? _addBtnWidth : 0;
+                var additionalWidth = HasAddButton && (_scrollListWidth <= _wrapperWidth) ? _addBtnWidth : 0;
 
                 // need to scroll tab bars
                 if (_activeTabElement.OffsetLeft + _activeTabElement.ClientWidth + additionalWidth > _scrollOffset + _wrapperWidth
@@ -772,7 +809,10 @@ namespace AntDesign
             }
 
             _invisibleTabs = _tabs.ToList();
-            _invisibleTabs.RemoveRange(invisibleHeadCount, visibleCount);
+            if (invisibleHeadCount + visibleCount <= _invisibleTabs.Count)
+            {
+                _invisibleTabs.RemoveRange(invisibleHeadCount, visibleCount);
+            }
         }
 
         private static int GetOverflowCount(decimal maxLength, decimal[] lengths, bool isRight = false)
